Struct aws_sdk_autoscaling::client::fluent_builders::CreateLaunchConfiguration

pub struct CreateLaunchConfiguration<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateLaunchConfiguration<C>

pub async fn send(
self
) -> Result<CreateLaunchConfigurationOutput, SdkError<CreateLaunchConfigurationError>> where
C: SmithyConnector,

pub fn launch_configuration_name(self, input: impl Into<String>) -> Self

The name of the launch configuration. This name must be unique per Region per account.

pub fn set_launch_configuration_name(self, input: Option<String>) -> Self

pub fn image_id(self, input: impl Into<String>) -> Self

The ID of the Amazon Machine Image (AMI) that was assigned during registration. For more information, see Finding an AMI in the Amazon EC2 User Guide for Linux Instances.

If you do not specify InstanceId, you must specify ImageId.

pub fn set_image_id(self, input: Option<String>) -> Self

pub fn key_name(self, input: impl Into<String>) -> Self

The name of the key pair. For more information, see Amazon EC2 Key Pairs in the Amazon EC2 User Guide for Linux Instances.

pub fn set_key_name(self, input: Option<String>) -> Self

pub fn security_groups(self, inp: impl Into<String>) -> Self

A list that contains the security groups to assign to the instances in the Auto Scaling group.

[EC2-VPC] Specify the security group IDs. For more information, see Security Groups for Your VPC in the Amazon Virtual Private Cloud User Guide.

[EC2-Classic] Specify either the security group names or the security group IDs. For more information, see Amazon EC2 Security Groups in the Amazon EC2 User Guide for Linux Instances.

pub fn set_security_groups(self, input: Option<Vec<String>>) -> Self

pub fn classic_link_vpc_id(self, input: impl Into<String>) -> Self

The ID of a ClassicLink-enabled VPC to link your EC2-Classic instances to. For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

This parameter can only be used if you are launching EC2-Classic instances.

pub fn set_classic_link_vpc_id(self, input: Option<String>) -> Self

pub fn classic_link_vpc_security_groups(self, inp: impl Into<String>) -> Self

The IDs of one or more security groups for the specified ClassicLink-enabled VPC. For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

If you specify the ClassicLinkVPCId parameter, you must specify this parameter.

pub fn set_classic_link_vpc_security_groups(
self,
input: Option<Vec<String>>
) -> Self

pub fn user_data(self, input: impl Into<String>) -> Self

The user data to make available to the launched EC2 instances. For more information, see Instance metadata and user data (Linux) and Instance metadata and user data (Windows). If you are using a command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text. User data is limited to 16 KB.

pub fn set_user_data(self, input: Option<String>) -> Self

pub fn instance_id(self, input: impl Into<String>) -> Self

The ID of the instance to use to create the launch configuration. The new launch configuration derives attributes from the instance, except for the block device mapping.

To create a launch configuration with a block device mapping or override any other instance attributes, specify them as part of the same request.

For more information, see Creating a launch configuration using an EC2 instance in the Amazon EC2 Auto Scaling User Guide.

If you do not specify InstanceId, you must specify both ImageId and InstanceType.

pub fn set_instance_id(self, input: Option<String>) -> Self

pub fn instance_type(self, input: impl Into<String>) -> Self

Specifies the instance type of the EC2 instance.

For information about available instance types, see Available Instance Types in the Amazon EC2 User Guide for Linux Instances.

If you do not specify InstanceId, you must specify InstanceType.

pub fn set_instance_type(self, input: Option<String>) -> Self

pub fn kernel_id(self, input: impl Into<String>) -> Self

The ID of the kernel associated with the AMI.

pub fn set_kernel_id(self, input: Option<String>) -> Self

pub fn ramdisk_id(self, input: impl Into<String>) -> Self

The ID of the RAM disk to select.

pub fn set_ramdisk_id(self, input: Option<String>) -> Self

pub fn block_device_mappings(self, inp: impl Into<BlockDeviceMapping>) -> Self

A block device mapping, which specifies the block devices for the instance. You can specify virtual devices and EBS volumes. For more information, see Block Device Mapping in the Amazon EC2 User Guide for Linux Instances.

pub fn set_block_device_mappings(
self,
input: Option<Vec<BlockDeviceMapping>>
) -> Self

pub fn instance_monitoring(self, input: InstanceMonitoring) -> Self

Controls whether instances in this group are launched with detailed (true) or basic (false) monitoring.

The default value is true (enabled).

When detailed monitoring is enabled, Amazon CloudWatch generates metrics every minute and your account is charged a fee. When you disable detailed monitoring, CloudWatch generates metrics every 5 minutes. For more information, see Configure Monitoring for Auto Scaling Instances in the Amazon EC2 Auto Scaling User Guide.

pub fn set_instance_monitoring(self, input: Option<InstanceMonitoring>) -> Self

pub fn spot_price(self, input: impl Into<String>) -> Self

The maximum hourly price to be paid for any Spot Instance launched to fulfill the request. Spot Instances are launched when the price you specify exceeds the current Spot price. For more information, see Requesting Spot Instances in the Amazon EC2 Auto Scaling User Guide.

When you change your maximum price by creating a new launch configuration, running instances will continue to run as long as the maximum price for those running instances is higher than the current Spot price.

pub fn set_spot_price(self, input: Option<String>) -> Self

pub fn iam_instance_profile(self, input: impl Into<String>) -> Self

The name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role.

For more information, see IAM role for applications that run on Amazon EC2 instances in the Amazon EC2 Auto Scaling User Guide.

pub fn set_iam_instance_profile(self, input: Option<String>) -> Self

pub fn ebs_optimized(self, input: bool) -> Self

Specifies whether the launch configuration is optimized for EBS I/O (true) or not (false). The optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization is not available with all instance types. Additional fees are incurred when you enable EBS optimization for an instance type that is not EBS-optimized by default. For more information, see Amazon EBS-Optimized Instances in the Amazon EC2 User Guide for Linux Instances.

The default value is false.

pub fn set_ebs_optimized(self, input: Option<bool>) -> Self

pub fn associate_public_ip_address(self, input: bool) -> Self

For Auto Scaling groups that are running in a virtual private cloud (VPC), specifies whether to assign a public IP address to the group's instances. If you specify true, each instance in the Auto Scaling group receives a unique public IP address. For more information, see Launching Auto Scaling instances in a VPC in the Amazon EC2 Auto Scaling User Guide.

If you specify this parameter, you must specify at least one subnet for VPCZoneIdentifier when you create your group.

If the instance is launched into a default subnet, the default is to assign a public IP address, unless you disabled the option to assign a public IP address on the subnet. If the instance is launched into a nondefault subnet, the default is not to assign a public IP address, unless you enabled the option to assign a public IP address on the subnet.

pub fn set_associate_public_ip_address(self, input: Option<bool>) -> Self

pub fn placement_tenancy(self, input: impl Into<String>) -> Self

The tenancy of the instance. An instance with dedicated tenancy runs on isolated, single-tenant hardware and can only be launched into a VPC.

To launch dedicated instances into a shared tenancy VPC (a VPC with the instance placement tenancy attribute set to default), you must set the value of this parameter to dedicated.

If you specify PlacementTenancy, you must specify at least one subnet for VPCZoneIdentifier when you create your group.

For more information, see Configuring instance tenancy with Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

Valid Values: default | dedicated

pub fn set_placement_tenancy(self, input: Option<String>) -> Self

pub fn metadata_options(self, input: InstanceMetadataOptions) -> Self

The metadata options for the instances. For more information, see Configuring the Instance Metadata Options in the Amazon EC2 Auto Scaling User Guide.

pub fn set_metadata_options(
self,
input: Option<InstanceMetadataOptions>
) -> Self