Struct aws_sdk_autoscaling::client::fluent_builders::PutLifecycleHook

pub struct PutLifecycleHook<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> PutLifecycleHook<C>

pub async fn send(
self
) -> Result<PutLifecycleHookOutput, SdkError<PutLifecycleHookError>> where
C: SmithyConnector,

pub fn lifecycle_hook_name(self, input: impl Into<String>) -> Self

The name of the lifecycle hook.

pub fn set_lifecycle_hook_name(self, input: Option<String>) -> Self

pub fn auto_scaling_group_name(self, input: impl Into<String>) -> Self

The name of the Auto Scaling group.

pub fn set_auto_scaling_group_name(self, input: Option<String>) -> Self

pub fn lifecycle_transition(self, input: impl Into<String>) -> Self

The instance state to which you want to attach the lifecycle hook. The valid values are:

autoscaling:EC2_INSTANCE_LAUNCHING

autoscaling:EC2_INSTANCE_TERMINATING

Required for new lifecycle hooks, but optional when updating existing hooks.

pub fn set_lifecycle_transition(self, input: Option<String>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target, for example, an Amazon SNS topic or an Amazon SQS queue.

Required for new lifecycle hooks, but optional when updating existing hooks.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn notification_target_arn(self, input: impl Into<String>) -> Self

The ARN of the notification target that Amazon EC2 Auto Scaling uses to notify you when an instance is in the transition state for the lifecycle hook. This target can be either an SQS queue or an SNS topic.

If you specify an empty string, this overrides the current ARN.

This operation uses the JSON format when sending notifications to an Amazon SQS queue, and an email key-value pair format when sending notifications to an Amazon SNS topic.

When you specify a notification target, Amazon EC2 Auto Scaling sends it a test message. Test messages contain the following additional key-value pair: "Event": "autoscaling:TEST_NOTIFICATION".

pub fn set_notification_target_arn(self, input: Option<String>) -> Self

pub fn notification_metadata(self, input: impl Into<String>) -> Self

Additional information that you want to include any time Amazon EC2 Auto Scaling sends a message to the notification target.

pub fn set_notification_metadata(self, input: Option<String>) -> Self

pub fn heartbeat_timeout(self, input: i32) -> Self

The maximum time, in seconds, that can elapse before the lifecycle hook times out. The range is from 30 to 7200 seconds. The default value is 3600 seconds (1 hour).

If the lifecycle hook times out, Amazon EC2 Auto Scaling performs the action that you specified in the DefaultResult parameter. You can prevent the lifecycle hook from timing out by calling the RecordLifecycleActionHeartbeat API.

pub fn set_heartbeat_timeout(self, input: Option<i32>) -> Self

pub fn default_result(self, input: impl Into<String>) -> Self

Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. This parameter can be either CONTINUE or ABANDON. The default value is ABANDON.

pub fn set_default_result(self, input: Option<String>) -> Self