Struct aws_sdk_autoscaling::client::fluent_builders::PutScalingPolicy

pub struct PutScalingPolicy<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> PutScalingPolicy<C>

pub async fn send(
self
) -> Result<PutScalingPolicyOutput, SdkError<PutScalingPolicyError>> where
C: SmithyConnector,

pub fn auto_scaling_group_name(self, input: impl Into<String>) -> Self

The name of the Auto Scaling group.

pub fn set_auto_scaling_group_name(self, input: Option<String>) -> Self

pub fn policy_name(self, input: impl Into<String>) -> Self

The name of the policy.

pub fn set_policy_name(self, input: Option<String>) -> Self

pub fn policy_type(self, input: impl Into<String>) -> Self

One of the following policy types:

TargetTrackingScaling

StepScaling

SimpleScaling (default)

PredictiveScaling

pub fn set_policy_type(self, input: Option<String>) -> Self

pub fn adjustment_type(self, input: impl Into<String>) -> Self

Specifies how the scaling adjustment is interpreted (for example, an absolute number or a percentage). The valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

Required if the policy type is StepScaling or SimpleScaling. For more information, see Scaling adjustment types in the Amazon EC2 Auto Scaling User Guide.

pub fn set_adjustment_type(self, input: Option<String>) -> Self

pub fn min_adjustment_step(self, input: i32) -> Self

Available for backward compatibility. Use MinAdjustmentMagnitude instead.

pub fn set_min_adjustment_step(self, input: Option<i32>) -> Self

pub fn min_adjustment_magnitude(self, input: i32) -> Self

The minimum value to scale by when the adjustment type is PercentChangeInCapacity. For example, suppose that you create a step scaling policy to scale out an Auto Scaling group by 25 percent and you specify a MinAdjustmentMagnitude of 2. If the group has 4 instances and the scaling policy is performed, 25 percent of 4 is 1. However, because you specified a MinAdjustmentMagnitude of 2, Amazon EC2 Auto Scaling scales out the group by 2 instances.

Valid only if the policy type is StepScaling or SimpleScaling. For more information, see Scaling adjustment types in the Amazon EC2 Auto Scaling User Guide.

Some Auto Scaling groups use instance weights. In this case, set the MinAdjustmentMagnitude to a value that is at least as large as your largest instance weight.

pub fn set_min_adjustment_magnitude(self, input: Option<i32>) -> Self

pub fn scaling_adjustment(self, input: i32) -> Self

The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity. For exact capacity, you must specify a positive value.

Required if the policy type is SimpleScaling. (Not used with any other policy type.)

pub fn set_scaling_adjustment(self, input: Option<i32>) -> Self

pub fn cooldown(self, input: i32) -> Self

The duration of the policy's cooldown period, in seconds. When a cooldown period is specified here, it overrides the default cooldown period defined for the Auto Scaling group.

Valid only if the policy type is SimpleScaling. For more information, see Scaling cooldowns for Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

pub fn set_cooldown(self, input: Option<i32>) -> Self

pub fn metric_aggregation_type(self, input: impl Into<String>) -> Self

The aggregation type for the CloudWatch metrics. The valid values are Minimum, Maximum, and Average. If the aggregation type is null, the value is treated as Average.

Valid only if the policy type is StepScaling.

pub fn set_metric_aggregation_type(self, input: Option<String>) -> Self

pub fn step_adjustments(self, inp: impl Into<StepAdjustment>) -> Self

A set of adjustments that enable you to scale based on the size of the alarm breach.

Required if the policy type is StepScaling. (Not used with any other policy type.)

pub fn set_step_adjustments(self, input: Option<Vec<StepAdjustment>>) -> Self

pub fn estimated_instance_warmup(self, input: i32) -> Self

The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics. If not provided, the default is to use the value from the default cooldown period for the Auto Scaling group.

Valid only if the policy type is TargetTrackingScaling or StepScaling.

pub fn set_estimated_instance_warmup(self, input: Option<i32>) -> Self

pub fn target_tracking_configuration(
self,
input: TargetTrackingConfiguration
) -> Self

A target tracking scaling policy. Provides support for predefined or customized metrics.

The following predefined metrics are available:

ASGAverageCPUUtilization

ASGAverageNetworkIn

ASGAverageNetworkOut

ALBRequestCountPerTarget

If you specify ALBRequestCountPerTarget for the metric, you must specify the ResourceLabel parameter with the PredefinedMetricSpecification.

For more information, see TargetTrackingConfiguration in the Amazon EC2 Auto Scaling API Reference.

Required if the policy type is TargetTrackingScaling.

pub fn set_target_tracking_configuration(
self,
input: Option<TargetTrackingConfiguration>
) -> Self

pub fn enabled(self, input: bool) -> Self

Indicates whether the scaling policy is enabled or disabled. The default is enabled. For more information, see Disabling a scaling policy for an Auto Scaling group in the Amazon EC2 Auto Scaling User Guide.

pub fn set_enabled(self, input: Option<bool>) -> Self

pub fn predictive_scaling_configuration(
self,
input: PredictiveScalingConfiguration
) -> Self

A predictive scaling policy. Provides support for only predefined metrics.

Predictive scaling works with CPU utilization, network in/out, and the Application Load Balancer request count.

For more information, see PredictiveScalingConfiguration in the Amazon EC2 Auto Scaling API Reference.

Required if the policy type is PredictiveScaling.

pub fn set_predictive_scaling_configuration(
self,
input: Option<PredictiveScalingConfiguration>
) -> Self