Module aws_sdk_autoscaling::error

Modules

active_instance_refresh_not_found_fault

See ActiveInstanceRefreshNotFoundFault

already_exists_fault

See AlreadyExistsFault

instance_refresh_in_progress_fault

See InstanceRefreshInProgressFault

invalid_next_token

See InvalidNextToken

limit_exceeded_fault

See LimitExceededFault

resource_contention_fault

See ResourceContentionFault

resource_in_use_fault

See ResourceInUseFault

scaling_activity_in_progress_fault

See ScalingActivityInProgressFault

service_linked_role_failure

See ServiceLinkedRoleFailure

Structs

ActiveInstanceRefreshNotFoundFault

The request failed because an active instance refresh for the specified Auto Scaling group was

AlreadyExistsFault

You already have an Auto Scaling group or launch configuration with this name.

AttachInstancesError

AttachLoadBalancerTargetGroupsError

AttachLoadBalancersError

BatchDeleteScheduledActionError

BatchPutScheduledUpdateGroupActionError

CancelInstanceRefreshError

CompleteLifecycleActionError

CreateAutoScalingGroupError

CreateLaunchConfigurationError

CreateOrUpdateTagsError

DeleteAutoScalingGroupError

DeleteLaunchConfigurationError

DeleteLifecycleHookError

DeleteNotificationConfigurationError

DeletePolicyError

DeleteScheduledActionError

DeleteTagsError

DeleteWarmPoolError

DescribeAccountLimitsError

DescribeAdjustmentTypesError

DescribeAutoScalingGroupsError

DescribeAutoScalingInstancesError

DescribeAutoScalingNotificationTypesError

DescribeInstanceRefreshesError

DescribeLaunchConfigurationsError

DescribeLifecycleHookTypesError

DescribeLifecycleHooksError

DescribeLoadBalancerTargetGroupsError

DescribeLoadBalancersError

DescribeMetricCollectionTypesError

DescribeNotificationConfigurationsError

DescribePoliciesError

DescribeScalingActivitiesError

DescribeScalingProcessTypesError

DescribeScheduledActionsError

DescribeTagsError

DescribeTerminationPolicyTypesError

DescribeWarmPoolError

DetachInstancesError

DetachLoadBalancerTargetGroupsError

DetachLoadBalancersError

DisableMetricsCollectionError

EnableMetricsCollectionError

EnterStandbyError

ExecutePolicyError

ExitStandbyError

GetPredictiveScalingForecastError

InstanceRefreshInProgressFault

The request failed because an active instance refresh operation already exists for the

InvalidNextToken

The NextToken value is not valid.

LimitExceededFault

You have already reached a limit for your Amazon EC2 Auto Scaling

PutLifecycleHookError

PutNotificationConfigurationError

PutScalingPolicyError

PutScheduledUpdateGroupActionError

PutWarmPoolError

RecordLifecycleActionHeartbeatError

ResourceContentionFault

You already have a pending update to an Amazon EC2 Auto Scaling resource (for example, an Auto Scaling group,

ResourceInUseFault

The operation can't be performed because the resource is in use.

ResumeProcessesError

ScalingActivityInProgressFault

The operation can't be performed because there are scaling activities in

ServiceLinkedRoleFailure

The service-linked role is not yet ready for use.

SetDesiredCapacityError

SetInstanceHealthError

SetInstanceProtectionError

StartInstanceRefreshError

SuspendProcessesError

TerminateInstanceInAutoScalingGroupError

UpdateAutoScalingGroupError

Enums

AttachInstancesErrorKind

AttachLoadBalancerTargetGroupsErrorKind

AttachLoadBalancersErrorKind

BatchDeleteScheduledActionErrorKind

BatchPutScheduledUpdateGroupActionErrorKind

CancelInstanceRefreshErrorKind

CompleteLifecycleActionErrorKind

CreateAutoScalingGroupErrorKind

CreateLaunchConfigurationErrorKind

CreateOrUpdateTagsErrorKind

DeleteAutoScalingGroupErrorKind

DeleteLaunchConfigurationErrorKind

DeleteLifecycleHookErrorKind

DeleteNotificationConfigurationErrorKind

DeletePolicyErrorKind

DeleteScheduledActionErrorKind

DeleteTagsErrorKind

DeleteWarmPoolErrorKind

DescribeAccountLimitsErrorKind

DescribeAdjustmentTypesErrorKind

DescribeAutoScalingGroupsErrorKind

DescribeAutoScalingInstancesErrorKind

DescribeAutoScalingNotificationTypesErrorKind

DescribeInstanceRefreshesErrorKind

DescribeLaunchConfigurationsErrorKind

DescribeLifecycleHookTypesErrorKind

DescribeLifecycleHooksErrorKind

DescribeLoadBalancerTargetGroupsErrorKind

DescribeLoadBalancersErrorKind

DescribeMetricCollectionTypesErrorKind

DescribeNotificationConfigurationsErrorKind

DescribePoliciesErrorKind

DescribeScalingActivitiesErrorKind

DescribeScalingProcessTypesErrorKind

DescribeScheduledActionsErrorKind

DescribeTagsErrorKind

DescribeTerminationPolicyTypesErrorKind

DescribeWarmPoolErrorKind

DetachInstancesErrorKind

DetachLoadBalancerTargetGroupsErrorKind

DetachLoadBalancersErrorKind

DisableMetricsCollectionErrorKind

EnableMetricsCollectionErrorKind

EnterStandbyErrorKind

ExecutePolicyErrorKind

ExitStandbyErrorKind

GetPredictiveScalingForecastErrorKind

PutLifecycleHookErrorKind

PutNotificationConfigurationErrorKind

PutScalingPolicyErrorKind

PutScheduledUpdateGroupActionErrorKind

PutWarmPoolErrorKind

RecordLifecycleActionHeartbeatErrorKind

ResumeProcessesErrorKind

SetDesiredCapacityErrorKind

SetInstanceHealthErrorKind

SetInstanceProtectionErrorKind

StartInstanceRefreshErrorKind

SuspendProcessesErrorKind

TerminateInstanceInAutoScalingGroupErrorKind

UpdateAutoScalingGroupErrorKind