Module aws_sdk_autoscaling::model

Modules

activity

See Activity

adjustment_type

See AdjustmentType

alarm

See Alarm

auto_scaling_group

See AutoScalingGroup

auto_scaling_instance_details

See AutoScalingInstanceDetails

block_device_mapping

See BlockDeviceMapping

capacity_forecast

See CapacityForecast

customized_metric_specification

See CustomizedMetricSpecification

ebs

See Ebs

enabled_metric

See EnabledMetric

failed_scheduled_update_group_action_request

See FailedScheduledUpdateGroupActionRequest

filter

See Filter

instance

See Instance

instance_metadata_options

See InstanceMetadataOptions

instance_monitoring

See InstanceMonitoring

instance_refresh

See InstanceRefresh

instance_refresh_live_pool_progress

See InstanceRefreshLivePoolProgress

instance_refresh_progress_details

See InstanceRefreshProgressDetails

instance_refresh_warm_pool_progress

See InstanceRefreshWarmPoolProgress

instances_distribution

See InstancesDistribution

launch_configuration

See LaunchConfiguration

launch_template

See LaunchTemplate

launch_template_overrides

See LaunchTemplateOverrides

launch_template_specification

See LaunchTemplateSpecification

lifecycle_hook

See LifecycleHook

lifecycle_hook_specification

See LifecycleHookSpecification

load_balancer_state

See LoadBalancerState

load_balancer_target_group_state

See LoadBalancerTargetGroupState

load_forecast

See LoadForecast

metric_collection_type

See MetricCollectionType

metric_dimension

See MetricDimension

metric_granularity_type

See MetricGranularityType

mixed_instances_policy

See MixedInstancesPolicy

notification_configuration

See NotificationConfiguration

predefined_metric_specification

See PredefinedMetricSpecification

predictive_scaling_configuration

See PredictiveScalingConfiguration

predictive_scaling_metric_specification

See PredictiveScalingMetricSpecification

predictive_scaling_predefined_load_metric

See PredictiveScalingPredefinedLoadMetric

predictive_scaling_predefined_metric_pair

See PredictiveScalingPredefinedMetricPair

predictive_scaling_predefined_scaling_metric

See PredictiveScalingPredefinedScalingMetric

process_type

See ProcessType

refresh_preferences

See RefreshPreferences

scaling_policy

See ScalingPolicy

scheduled_update_group_action

See ScheduledUpdateGroupAction

scheduled_update_group_action_request

See ScheduledUpdateGroupActionRequest

step_adjustment

See StepAdjustment

suspended_process

See SuspendedProcess

tag

See Tag

tag_description

See TagDescription

target_tracking_configuration

See TargetTrackingConfiguration

warm_pool_configuration

See WarmPoolConfiguration

Structs

Activity

Describes scaling activity, which is a long-running process that represents a change

AdjustmentType

Describes a policy adjustment type.

Alarm

Describes an alarm.

AutoScalingGroup

Describes an Auto Scaling group.

AutoScalingInstanceDetails

Describes an EC2 instance associated with an Auto Scaling group.

BlockDeviceMapping

Describes a block device mapping.

CapacityForecast

A GetPredictiveScalingForecast call returns the capacity forecast for a

CustomizedMetricSpecification

Represents a CloudWatch metric of your choosing for a target tracking scaling policy to use

Ebs

Describes information used to set up an Amazon EBS volume specified in a block device

EnabledMetric

Describes an enabled metric.

FailedScheduledUpdateGroupActionRequest

Describes a scheduled action that could not be created, updated, or deleted.

Filter

Describes a filter that is used to return a more specific list of results when

Instance

Describes an EC2 instance.

InstanceMetadataOptions

The metadata options for the instances. For more information, see Configuring the Instance Metadata Options in the

InstanceMonitoring

Describes whether detailed monitoring is enabled for the Auto Scaling instances.

InstanceRefresh

Describes an instance refresh for an Auto Scaling group.

InstanceRefreshLivePoolProgress

Reports the progress of an instance refresh on instances that are in the Auto Scaling

InstanceRefreshProgressDetails

Reports the progress of an instance refresh on an Auto Scaling group that has a warm pool.

InstanceRefreshWarmPoolProgress

Reports the progress of an instance refresh on instances that are in the warm

InstancesDistribution

Describes an instances distribution for an Auto Scaling group with a MixedInstancesPolicy.

LaunchConfiguration

Describes a launch configuration.

LaunchTemplate

Describes a launch template and overrides.

LaunchTemplateOverrides

Describes an override for a launch template. The maximum number of instance types that

LaunchTemplateSpecification

Describes the Amazon EC2 launch template and the launch template version that can be used

LifecycleHook

Describes a lifecycle hook, which tells Amazon EC2 Auto Scaling that you want to perform an action

LifecycleHookSpecification

Describes information used to specify a lifecycle hook for an Auto Scaling

LoadBalancerState

Describes the state of a Classic Load Balancer.

LoadBalancerTargetGroupState

Describes the state of a target group.

LoadForecast

A GetPredictiveScalingForecast call returns the load forecast for a

MetricCollectionType

Describes a metric.

MetricDimension

Describes the dimension of a metric.

MetricGranularityType

Describes a granularity of a metric.

MixedInstancesPolicy

Describes a mixed instances policy for an Auto Scaling group. With mixed instances, your Auto Scaling

NotificationConfiguration

Describes a notification.

PredefinedMetricSpecification

Represents a predefined metric for a target tracking scaling policy to use with

PredictiveScalingConfiguration

Represents a predictive scaling policy configuration to use with Amazon EC2 Auto Scaling.

PredictiveScalingMetricSpecification

This structure specifies the metrics and target utilization settings for a predictive

PredictiveScalingPredefinedLoadMetric

Describes a load metric for a predictive scaling policy.

PredictiveScalingPredefinedMetricPair

Represents a metric pair for a predictive scaling policy.

PredictiveScalingPredefinedScalingMetric

Describes a scaling metric for a predictive scaling policy.

ProcessType

Describes a process type.

RefreshPreferences

Describes information used to start an instance refresh.

ScalingPolicy

Describes a scaling policy.

ScheduledUpdateGroupAction

Describes a scheduled scaling

ScheduledUpdateGroupActionRequest

Describes information used for one or more scheduled scaling action updates in a BatchPutScheduledUpdateGroupAction operation.

StepAdjustment

Describes information used to create a step adjustment for a step scaling

SuspendedProcess

Describes an auto scaling process that has been suspended.

Tag

Describes a tag for an Auto Scaling group.

TagDescription

Describes a tag for an Auto Scaling group.

TargetTrackingConfiguration

Represents a target tracking scaling policy configuration to use with Amazon EC2 Auto Scaling.

WarmPoolConfiguration

Describes a warm pool configuration.

Enums

InstanceMetadataEndpointState

InstanceMetadataHttpTokensState

InstanceRefreshStatus

LifecycleState

MetricStatistic

MetricType

PredefinedLoadMetricType

PredefinedMetricPairType

PredefinedScalingMetricType

PredictiveScalingMaxCapacityBreachBehavior

PredictiveScalingMode

RefreshStrategy

ScalingActivityStatusCode

WarmPoolState

WarmPoolStatus