Struct aws_sdk_autoscaling::model::auto_scaling_group::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AutoScalingGroup

Implementations

impl Builder

pub fn auto_scaling_group_name(self, input: impl Into<String>) -> Self

The name of the Auto Scaling group.

pub fn set_auto_scaling_group_name(self, input: Option<String>) -> Self

pub fn auto_scaling_group_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Auto Scaling group.

pub fn set_auto_scaling_group_arn(self, input: Option<String>) -> Self

pub fn launch_configuration_name(self, input: impl Into<String>) -> Self

The name of the associated launch configuration.

pub fn set_launch_configuration_name(self, input: Option<String>) -> Self

pub fn launch_template(self, input: LaunchTemplateSpecification) -> Self

The launch template for the group.

pub fn set_launch_template(
self,
input: Option<LaunchTemplateSpecification>
) -> Self

pub fn mixed_instances_policy(self, input: MixedInstancesPolicy) -> Self

The mixed instances policy for the group.

pub fn set_mixed_instances_policy(
self,
input: Option<MixedInstancesPolicy>
) -> Self

pub fn min_size(self, input: i32) -> Self

The minimum size of the group.

pub fn set_min_size(self, input: Option<i32>) -> Self

pub fn max_size(self, input: i32) -> Self

The maximum size of the group.

pub fn set_max_size(self, input: Option<i32>) -> Self

pub fn desired_capacity(self, input: i32) -> Self

The desired size of the group.

pub fn set_desired_capacity(self, input: Option<i32>) -> Self

pub fn predicted_capacity(self, input: i32) -> Self

The predicted capacity of the group when it has a predictive scaling policy.

pub fn set_predicted_capacity(self, input: Option<i32>) -> Self

pub fn default_cooldown(self, input: i32) -> Self

The duration of the default cooldown period, in seconds.

pub fn set_default_cooldown(self, input: Option<i32>) -> Self

pub fn availability_zones(self, input: impl Into<String>) -> Self

pub fn set_availability_zones(self, input: Option<Vec<String>>) -> Self

pub fn load_balancer_names(self, input: impl Into<String>) -> Self

pub fn set_load_balancer_names(self, input: Option<Vec<String>>) -> Self

pub fn target_group_ar_ns(self, input: impl Into<String>) -> Self

pub fn set_target_group_ar_ns(self, input: Option<Vec<String>>) -> Self

pub fn health_check_type(self, input: impl Into<String>) -> Self

The service to use for the health checks. The valid values are EC2 and ELB. If you configure an Auto Scaling group to use ELB health checks, it considers the instance unhealthy if it fails either the EC2 status checks or the load balancer health checks.

pub fn set_health_check_type(self, input: Option<String>) -> Self

pub fn health_check_grace_period(self, input: i32) -> Self

The amount of time, in seconds, that Amazon EC2 Auto Scaling waits before checking the health status of an EC2 instance that has come into service.

pub fn set_health_check_grace_period(self, input: Option<i32>) -> Self

pub fn instances(self, input: impl Into<Instance>) -> Self

pub fn set_instances(self, input: Option<Vec<Instance>>) -> Self

pub fn created_time(self, input: Instant) -> Self

The date and time the group was created.

pub fn set_created_time(self, input: Option<Instant>) -> Self

pub fn suspended_processes(self, input: impl Into<SuspendedProcess>) -> Self

pub fn set_suspended_processes(
self,
input: Option<Vec<SuspendedProcess>>
) -> Self

pub fn placement_group(self, input: impl Into<String>) -> Self

The name of the placement group into which to launch your instances, if any.

pub fn set_placement_group(self, input: Option<String>) -> Self

pub fn vpc_zone_identifier(self, input: impl Into<String>) -> Self

One or more subnet IDs, if applicable, separated by commas.

pub fn set_vpc_zone_identifier(self, input: Option<String>) -> Self

pub fn enabled_metrics(self, input: impl Into<EnabledMetric>) -> Self

pub fn set_enabled_metrics(self, input: Option<Vec<EnabledMetric>>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

The current state of the group when the DeleteAutoScalingGroup operation is in progress.

pub fn set_status(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<TagDescription>) -> Self

pub fn set_tags(self, input: Option<Vec<TagDescription>>) -> Self

pub fn termination_policies(self, input: impl Into<String>) -> Self

pub fn set_termination_policies(self, input: Option<Vec<String>>) -> Self

pub fn new_instances_protected_from_scale_in(self, input: bool) -> Self

Indicates whether newly launched instances are protected from termination by Amazon EC2 Auto Scaling when scaling in.

pub fn set_new_instances_protected_from_scale_in(
self,
input: Option<bool>
) -> Self

pub fn service_linked_role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the service-linked role that the Auto Scaling group uses to call other Amazon Web Services on your behalf.

pub fn set_service_linked_role_arn(self, input: Option<String>) -> Self

pub fn max_instance_lifetime(self, input: i32) -> Self

The maximum amount of time, in seconds, that an instance can be in service.

Valid Range: Minimum value of 0.

pub fn set_max_instance_lifetime(self, input: Option<i32>) -> Self

pub fn capacity_rebalance(self, input: bool) -> Self

Indicates whether Capacity Rebalancing is enabled.

pub fn set_capacity_rebalance(self, input: Option<bool>) -> Self

pub fn warm_pool_configuration(self, input: WarmPoolConfiguration) -> Self

The warm pool for the group.

pub fn set_warm_pool_configuration(
self,
input: Option<WarmPoolConfiguration>
) -> Self

pub fn warm_pool_size(self, input: i32) -> Self

The current size of the warm pool.

pub fn set_warm_pool_size(self, input: Option<i32>) -> Self

pub fn context(self, input: impl Into<String>) -> Self

Reserved.

pub fn set_context(self, input: Option<String>) -> Self

pub fn build(self) -> AutoScalingGroup

Consumes the builder and constructs a AutoScalingGroup