Struct aws_sdk_autoscaling::model::instance_requirements::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for InstanceRequirements

Implementations

impl Builder

pub fn v_cpu_count(self, input: VCpuCountRequest) -> Self

The minimum and maximum number of vCPUs for an instance type.

pub fn set_v_cpu_count(self, input: Option<VCpuCountRequest>) -> Self

The minimum and maximum number of vCPUs for an instance type.

pub fn memory_mi_b(self, input: MemoryMiBRequest) -> Self

The minimum and maximum instance memory size for an instance type, in MiB.

pub fn set_memory_mi_b(self, input: Option<MemoryMiBRequest>) -> Self

The minimum and maximum instance memory size for an instance type, in MiB.

pub fn cpu_manufacturers(self, input: impl Into<CpuManufacturer>) -> Self

Appends an item to cpu_manufacturers.

To override the contents of this collection use set_cpu_manufacturers.

Lists which specific CPU manufacturers to include.

For instance types with Intel CPUs, specify intel.

For instance types with AMD CPUs, specify amd.

For instance types with Amazon Web Services CPUs, specify amazon-web-services.

Don't confuse the CPU hardware manufacturer with the CPU hardware architecture. Instances will be launched with a compatible CPU architecture based on the Amazon Machine Image (AMI) that you specify in your launch template.

Default: Any manufacturer

pub fn set_cpu_manufacturers(self, input: Option<Vec<CpuManufacturer>>) -> Self

Lists which specific CPU manufacturers to include.

For instance types with Intel CPUs, specify intel.

For instance types with AMD CPUs, specify amd.

For instance types with Amazon Web Services CPUs, specify amazon-web-services.

Don't confuse the CPU hardware manufacturer with the CPU hardware architecture. Instances will be launched with a compatible CPU architecture based on the Amazon Machine Image (AMI) that you specify in your launch template.

Default: Any manufacturer

pub fn memory_gi_b_per_v_cpu(self, input: MemoryGiBPerVCpuRequest) -> Self

The minimum and maximum amount of memory per vCPU for an instance type, in GiB.

Default: No minimum or maximum

pub fn set_memory_gi_b_per_v_cpu(
self,
input: Option<MemoryGiBPerVCpuRequest>
) -> Self

The minimum and maximum amount of memory per vCPU for an instance type, in GiB.

Default: No minimum or maximum

pub fn excluded_instance_types(self, input: impl Into<String>) -> Self

Appends an item to excluded_instance_types.

To override the contents of this collection use set_excluded_instance_types.

Lists which instance types to exclude. You can use strings with one or more wild cards, represented by an asterisk (*). The following are examples: c5*, m5a.*, r*, *3*.

For example, if you specify c5*, you are excluding the entire C5 instance family, which includes all C5a and C5n instance types. If you specify m5a.*, you are excluding all the M5a instance types, but not the M5n instance types.

Default: No excluded instance types

pub fn set_excluded_instance_types(self, input: Option<Vec<String>>) -> Self

Lists which instance types to exclude. You can use strings with one or more wild cards, represented by an asterisk (*). The following are examples: c5*, m5a.*, r*, *3*.

For example, if you specify c5*, you are excluding the entire C5 instance family, which includes all C5a and C5n instance types. If you specify m5a.*, you are excluding all the M5a instance types, but not the M5n instance types.

Default: No excluded instance types

pub fn instance_generations(self, input: impl Into<InstanceGeneration>) -> Self

Appends an item to instance_generations.

To override the contents of this collection use set_instance_generations.

Indicates whether current or previous generation instance types are included.

For current generation instance types, specify current. The current generation includes EC2 instance types currently recommended for use. This typically includes the latest two to three generations in each instance family. For more information, see Instance types in the Amazon EC2 User Guide for Linux Instances.

For previous generation instance types, specify previous.

Default: Any current or previous generation

pub fn set_instance_generations(
self,
input: Option<Vec<InstanceGeneration>>
) -> Self

Indicates whether current or previous generation instance types are included.

For current generation instance types, specify current. The current generation includes EC2 instance types currently recommended for use. This typically includes the latest two to three generations in each instance family. For more information, see Instance types in the Amazon EC2 User Guide for Linux Instances.

For previous generation instance types, specify previous.

Default: Any current or previous generation

pub fn spot_max_price_percentage_over_lowest_price(self, input: i32) -> Self

The price protection threshold for Spot Instances. This is the maximum you’ll pay for a Spot Instance, expressed as a percentage higher than the cheapest M, C, or R instance type with your specified attributes. When Amazon EC2 Auto Scaling selects instance types with your attributes, we will exclude instance types whose price is higher than your threshold. The parameter accepts an integer, which Amazon EC2 Auto Scaling interprets as a percentage. To turn off price protection, specify a high value, such as 999999.

Default: 100

pub fn set_spot_max_price_percentage_over_lowest_price(
self,
input: Option<i32>
) -> Self

The price protection threshold for Spot Instances. This is the maximum you’ll pay for a Spot Instance, expressed as a percentage higher than the cheapest M, C, or R instance type with your specified attributes. When Amazon EC2 Auto Scaling selects instance types with your attributes, we will exclude instance types whose price is higher than your threshold. The parameter accepts an integer, which Amazon EC2 Auto Scaling interprets as a percentage. To turn off price protection, specify a high value, such as 999999.

Default: 100

pub fn on_demand_max_price_percentage_over_lowest_price(
self,
input: i32
) -> Self

The price protection threshold for On-Demand Instances. This is the maximum you’ll pay for an On-Demand Instance, expressed as a percentage higher than the cheapest M, C, or R instance type with your specified attributes. When Amazon EC2 Auto Scaling selects instance types with your attributes, we will exclude instance types whose price is higher than your threshold. The parameter accepts an integer, which Amazon EC2 Auto Scaling interprets as a percentage. To turn off price protection, specify a high value, such as 999999.

Default: 20

pub fn set_on_demand_max_price_percentage_over_lowest_price(
self,
input: Option<i32>
) -> Self

The price protection threshold for On-Demand Instances. This is the maximum you’ll pay for an On-Demand Instance, expressed as a percentage higher than the cheapest M, C, or R instance type with your specified attributes. When Amazon EC2 Auto Scaling selects instance types with your attributes, we will exclude instance types whose price is higher than your threshold. The parameter accepts an integer, which Amazon EC2 Auto Scaling interprets as a percentage. To turn off price protection, specify a high value, such as 999999.

Default: 20

pub fn bare_metal(self, input: BareMetal) -> Self

Indicates whether bare metal instance types are included, excluded, or required.

Default: excluded

pub fn set_bare_metal(self, input: Option<BareMetal>) -> Self

Indicates whether bare metal instance types are included, excluded, or required.

Default: excluded

pub fn burstable_performance(self, input: BurstablePerformance) -> Self

Indicates whether burstable performance instance types are included, excluded, or required. For more information, see Burstable performance instances in the Amazon EC2 User Guide for Linux Instances.

Default: excluded

pub fn set_burstable_performance(
self,
input: Option<BurstablePerformance>
) -> Self

Indicates whether burstable performance instance types are included, excluded, or required. For more information, see Burstable performance instances in the Amazon EC2 User Guide for Linux Instances.

Default: excluded

pub fn require_hibernate_support(self, input: bool) -> Self

Indicates whether instance types must provide On-Demand Instance hibernation support.

Default: false

pub fn set_require_hibernate_support(self, input: Option<bool>) -> Self

Indicates whether instance types must provide On-Demand Instance hibernation support.

Default: false

pub fn network_interface_count(
self,
input: NetworkInterfaceCountRequest
) -> Self

The minimum and maximum number of network interfaces for an instance type.

Default: No minimum or maximum

pub fn set_network_interface_count(
self,
input: Option<NetworkInterfaceCountRequest>
) -> Self

The minimum and maximum number of network interfaces for an instance type.

Default: No minimum or maximum

pub fn local_storage(self, input: LocalStorage) -> Self

Indicates whether instance types with instance store volumes are included, excluded, or required. For more information, see Amazon EC2 instance store in the Amazon EC2 User Guide for Linux Instances.

Default: included

pub fn set_local_storage(self, input: Option<LocalStorage>) -> Self

Indicates whether instance types with instance store volumes are included, excluded, or required. For more information, see Amazon EC2 instance store in the Amazon EC2 User Guide for Linux Instances.

Default: included

pub fn local_storage_types(self, input: impl Into<LocalStorageType>) -> Self

Appends an item to local_storage_types.

To override the contents of this collection use set_local_storage_types.

Indicates the type of local storage that is required.

For instance types with hard disk drive (HDD) storage, specify hdd.

For instance types with solid state drive (SSD) storage, specify sdd.

Default: Any local storage type

pub fn set_local_storage_types(
self,
input: Option<Vec<LocalStorageType>>
) -> Self

Indicates the type of local storage that is required.

For instance types with hard disk drive (HDD) storage, specify hdd.

For instance types with solid state drive (SSD) storage, specify sdd.

Default: Any local storage type

pub fn total_local_storage_gb(self, input: TotalLocalStorageGbRequest) -> Self

The minimum and maximum total local storage size for an instance type, in GB.

Default: No minimum or maximum

pub fn set_total_local_storage_gb(
self,
input: Option<TotalLocalStorageGbRequest>
) -> Self

The minimum and maximum total local storage size for an instance type, in GB.

Default: No minimum or maximum

pub fn baseline_ebs_bandwidth_mbps(
self,
input: BaselineEbsBandwidthMbpsRequest
) -> Self

The minimum and maximum baseline bandwidth performance for an instance type, in Mbps. For more information, see Amazon EBS–optimized instances in the Amazon EC2 User Guide for Linux Instances.

Default: No minimum or maximum

pub fn set_baseline_ebs_bandwidth_mbps(
self,
input: Option<BaselineEbsBandwidthMbpsRequest>
) -> Self

The minimum and maximum baseline bandwidth performance for an instance type, in Mbps. For more information, see Amazon EBS–optimized instances in the Amazon EC2 User Guide for Linux Instances.

Default: No minimum or maximum

pub fn accelerator_types(self, input: impl Into<AcceleratorType>) -> Self

Appends an item to accelerator_types.

To override the contents of this collection use set_accelerator_types.

Lists the accelerator types that must be on an instance type.

For instance types with GPU accelerators, specify gpu.

For instance types with FPGA accelerators, specify fpga.

For instance types with inference accelerators, specify inference.

Default: Any accelerator type

pub fn set_accelerator_types(self, input: Option<Vec<AcceleratorType>>) -> Self

Lists the accelerator types that must be on an instance type.

For instance types with GPU accelerators, specify gpu.

For instance types with FPGA accelerators, specify fpga.

For instance types with inference accelerators, specify inference.

Default: Any accelerator type

pub fn accelerator_count(self, input: AcceleratorCountRequest) -> Self

The minimum and maximum number of accelerators (GPUs, FPGAs, or Amazon Web Services Inferentia chips) for an instance type.

To exclude accelerator-enabled instance types, set Max to 0.

Default: No minimum or maximum

pub fn set_accelerator_count(
self,
input: Option<AcceleratorCountRequest>
) -> Self

The minimum and maximum number of accelerators (GPUs, FPGAs, or Amazon Web Services Inferentia chips) for an instance type.

To exclude accelerator-enabled instance types, set Max to 0.

Default: No minimum or maximum

pub fn accelerator_manufacturers(
self,
input: impl Into<AcceleratorManufacturer>
) -> Self

Appends an item to accelerator_manufacturers.

To override the contents of this collection use set_accelerator_manufacturers.

Indicates whether instance types must have accelerators by specific manufacturers.

For instance types with NVIDIA devices, specify nvidia.

For instance types with AMD devices, specify amd.

For instance types with Amazon Web Services devices, specify amazon-web-services.

For instance types with Xilinx devices, specify xilinx.

Default: Any manufacturer

pub fn set_accelerator_manufacturers(
self,
input: Option<Vec<AcceleratorManufacturer>>
) -> Self

Indicates whether instance types must have accelerators by specific manufacturers.

For instance types with NVIDIA devices, specify nvidia.

For instance types with AMD devices, specify amd.

For instance types with Amazon Web Services devices, specify amazon-web-services.

For instance types with Xilinx devices, specify xilinx.

Default: Any manufacturer

pub fn accelerator_names(self, input: impl Into<AcceleratorName>) -> Self

Appends an item to accelerator_names.

To override the contents of this collection use set_accelerator_names.

Lists the accelerators that must be on an instance type.

For instance types with NVIDIA A100 GPUs, specify a100.

For instance types with NVIDIA V100 GPUs, specify v100.

For instance types with NVIDIA K80 GPUs, specify k80.

For instance types with NVIDIA T4 GPUs, specify t4.

For instance types with NVIDIA M60 GPUs, specify m60.

For instance types with AMD Radeon Pro V520 GPUs, specify radeon-pro-v520.

For instance types with Xilinx VU9P FPGAs, specify vu9p.

Default: Any accelerator

pub fn set_accelerator_names(self, input: Option<Vec<AcceleratorName>>) -> Self

Lists the accelerators that must be on an instance type.

For instance types with NVIDIA A100 GPUs, specify a100.

For instance types with NVIDIA V100 GPUs, specify v100.

For instance types with NVIDIA K80 GPUs, specify k80.

For instance types with NVIDIA T4 GPUs, specify t4.

For instance types with NVIDIA M60 GPUs, specify m60.

For instance types with AMD Radeon Pro V520 GPUs, specify radeon-pro-v520.

For instance types with Xilinx VU9P FPGAs, specify vu9p.

Default: Any accelerator

pub fn accelerator_total_memory_mi_b(
self,
input: AcceleratorTotalMemoryMiBRequest
) -> Self

The minimum and maximum total memory size for the accelerators on an instance type, in MiB.

Default: No minimum or maximum

pub fn set_accelerator_total_memory_mi_b(
self,
input: Option<AcceleratorTotalMemoryMiBRequest>
) -> Self

The minimum and maximum total memory size for the accelerators on an instance type, in MiB.

Default: No minimum or maximum

pub fn build(self) -> InstanceRequirements

Consumes the builder and constructs a InstanceRequirements