Struct aws_sdk_autoscaling::model::launch_configuration::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for LaunchConfiguration

Implementations

impl Builder

pub fn launch_configuration_name(self, input: impl Into<String>) -> Self

The name of the launch configuration.

pub fn set_launch_configuration_name(self, input: Option<String>) -> Self

pub fn launch_configuration_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the launch configuration.

pub fn set_launch_configuration_arn(self, input: Option<String>) -> Self

pub fn image_id(self, input: impl Into<String>) -> Self

The ID of the Amazon Machine Image (AMI) to use to launch your EC2 instances. For more information, see Finding an AMI in the Amazon EC2 User Guide for Linux Instances.

pub fn set_image_id(self, input: Option<String>) -> Self

pub fn key_name(self, input: impl Into<String>) -> Self

The name of the key pair.

For more information, see Amazon EC2 Key Pairs in the Amazon EC2 User Guide for Linux Instances.

pub fn set_key_name(self, input: Option<String>) -> Self

pub fn security_groups(self, input: impl Into<String>) -> Self

pub fn set_security_groups(self, input: Option<Vec<String>>) -> Self

pub fn classic_link_vpc_id(self, input: impl Into<String>) -> Self

The ID of a ClassicLink-enabled VPC to link your EC2-Classic instances to. For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

pub fn set_classic_link_vpc_id(self, input: Option<String>) -> Self

pub fn classic_link_vpc_security_groups(self, input: impl Into<String>) -> Self

pub fn set_classic_link_vpc_security_groups(
self,
input: Option<Vec<String>>
) -> Self

pub fn user_data(self, input: impl Into<String>) -> Self

The user data to make available to the launched EC2 instances. For more information, see Instance metadata and user data (Linux) and Instance metadata and user data (Windows). If you are using a command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text. User data is limited to 16 KB.

pub fn set_user_data(self, input: Option<String>) -> Self

pub fn instance_type(self, input: impl Into<String>) -> Self

The instance type for the instances.

For information about available instance types, see Available Instance Types in the Amazon EC2 User Guide for Linux Instances.

pub fn set_instance_type(self, input: Option<String>) -> Self

pub fn kernel_id(self, input: impl Into<String>) -> Self

The ID of the kernel associated with the AMI.

pub fn set_kernel_id(self, input: Option<String>) -> Self

pub fn ramdisk_id(self, input: impl Into<String>) -> Self

The ID of the RAM disk associated with the AMI.

pub fn set_ramdisk_id(self, input: Option<String>) -> Self

pub fn block_device_mappings(self, input: impl Into<BlockDeviceMapping>) -> Self

pub fn set_block_device_mappings(
self,
input: Option<Vec<BlockDeviceMapping>>
) -> Self

pub fn instance_monitoring(self, input: InstanceMonitoring) -> Self

Controls whether instances in this group are launched with detailed (true) or basic (false) monitoring.

For more information, see Configure Monitoring for Auto Scaling Instances in the Amazon EC2 Auto Scaling User Guide.

pub fn set_instance_monitoring(self, input: Option<InstanceMonitoring>) -> Self

pub fn spot_price(self, input: impl Into<String>) -> Self

The maximum hourly price to be paid for any Spot Instance launched to fulfill the request. Spot Instances are launched when the price you specify exceeds the current Spot price. For more information, see Requesting Spot Instances in the Amazon EC2 Auto Scaling User Guide.

pub fn set_spot_price(self, input: Option<String>) -> Self

pub fn iam_instance_profile(self, input: impl Into<String>) -> Self

The name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role. For more information, see IAM role for applications that run on Amazon EC2 instances in the Amazon EC2 Auto Scaling User Guide.

pub fn set_iam_instance_profile(self, input: Option<String>) -> Self

pub fn created_time(self, input: Instant) -> Self

The creation date and time for the launch configuration.

pub fn set_created_time(self, input: Option<Instant>) -> Self

pub fn ebs_optimized(self, input: bool) -> Self

Specifies whether the launch configuration is optimized for EBS I/O (true) or not (false). For more information, see Amazon EBS-Optimized Instances in the Amazon EC2 User Guide for Linux Instances.

pub fn set_ebs_optimized(self, input: Option<bool>) -> Self

pub fn associate_public_ip_address(self, input: bool) -> Self

For Auto Scaling groups that are running in a VPC, specifies whether to assign a public IP address to the group's instances. For more information, see Launching Auto Scaling instances in a VPC in the Amazon EC2 Auto Scaling User Guide.

pub fn set_associate_public_ip_address(self, input: Option<bool>) -> Self

pub fn placement_tenancy(self, input: impl Into<String>) -> Self

The tenancy of the instance, either default or dedicated. An instance with dedicated tenancy runs on isolated, single-tenant hardware and can only be launched into a VPC.

For more information, see Configuring instance tenancy with Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

pub fn set_placement_tenancy(self, input: Option<String>) -> Self

pub fn metadata_options(self, input: InstanceMetadataOptions) -> Self

The metadata options for the instances. For more information, see Configuring the Instance Metadata Options in the Amazon EC2 Auto Scaling User Guide.

pub fn set_metadata_options(
self,
input: Option<InstanceMetadataOptions>
) -> Self

pub fn build(self) -> LaunchConfiguration

Consumes the builder and constructs a LaunchConfiguration