Struct aws_sdk_autoscaling::model::scaling_policy::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ScalingPolicy

Implementations

impl Builder

pub fn auto_scaling_group_name(self, input: impl Into<String>) -> Self

The name of the Auto Scaling group.

pub fn set_auto_scaling_group_name(self, input: Option<String>) -> Self

pub fn policy_name(self, input: impl Into<String>) -> Self

The name of the scaling policy.

pub fn set_policy_name(self, input: Option<String>) -> Self

pub fn policy_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the policy.

pub fn set_policy_arn(self, input: Option<String>) -> Self

pub fn policy_type(self, input: impl Into<String>) -> Self

One of the following policy types:

TargetTrackingScaling

StepScaling

SimpleScaling (default)

PredictiveScaling

For more information, see Target tracking scaling policies and Step and simple scaling policies in the Amazon EC2 Auto Scaling User Guide.

pub fn set_policy_type(self, input: Option<String>) -> Self

pub fn adjustment_type(self, input: impl Into<String>) -> Self

Specifies how the scaling adjustment is interpreted (for example, an absolute number or a percentage). The valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

pub fn set_adjustment_type(self, input: Option<String>) -> Self

pub fn min_adjustment_step(self, input: i32) -> Self

Available for backward compatibility. Use MinAdjustmentMagnitude instead.

pub fn set_min_adjustment_step(self, input: Option<i32>) -> Self

pub fn min_adjustment_magnitude(self, input: i32) -> Self

The minimum value to scale by when the adjustment type is PercentChangeInCapacity.

pub fn set_min_adjustment_magnitude(self, input: Option<i32>) -> Self

pub fn scaling_adjustment(self, input: i32) -> Self

The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity.

pub fn set_scaling_adjustment(self, input: Option<i32>) -> Self

pub fn cooldown(self, input: i32) -> Self

The duration of the policy's cooldown period, in seconds.

pub fn set_cooldown(self, input: Option<i32>) -> Self

pub fn step_adjustments(self, input: impl Into<StepAdjustment>) -> Self

pub fn set_step_adjustments(self, input: Option<Vec<StepAdjustment>>) -> Self

pub fn metric_aggregation_type(self, input: impl Into<String>) -> Self

The aggregation type for the CloudWatch metrics. The valid values are Minimum, Maximum, and Average.

pub fn set_metric_aggregation_type(self, input: Option<String>) -> Self

pub fn estimated_instance_warmup(self, input: i32) -> Self

The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics.

pub fn set_estimated_instance_warmup(self, input: Option<i32>) -> Self

pub fn alarms(self, input: impl Into<Alarm>) -> Self

pub fn set_alarms(self, input: Option<Vec<Alarm>>) -> Self

pub fn target_tracking_configuration(
self,
input: TargetTrackingConfiguration
) -> Self

A target tracking scaling policy.

pub fn set_target_tracking_configuration(
self,
input: Option<TargetTrackingConfiguration>
) -> Self

pub fn enabled(self, input: bool) -> Self

Indicates whether the policy is enabled (true) or disabled (false).

pub fn set_enabled(self, input: Option<bool>) -> Self

pub fn predictive_scaling_configuration(
self,
input: PredictiveScalingConfiguration
) -> Self

A predictive scaling policy.

pub fn set_predictive_scaling_configuration(
self,
input: Option<PredictiveScalingConfiguration>
) -> Self

pub fn build(self) -> ScalingPolicy

Consumes the builder and constructs a ScalingPolicy