Struct aws_sdk_autoscaling::model::InstanceRequirements

#[non_exhaustive]
pub struct InstanceRequirements {

    pub v_cpu_count: Option<VCpuCountRequest>,
    pub memory_mi_b: Option<MemoryMiBRequest>,
    pub cpu_manufacturers: Option<Vec<CpuManufacturer>>,
    pub memory_gi_b_per_v_cpu: Option<MemoryGiBPerVCpuRequest>,
    pub excluded_instance_types: Option<Vec<String>>,
    pub instance_generations: Option<Vec<InstanceGeneration>>,
    pub spot_max_price_percentage_over_lowest_price: Option<i32>,
    pub on_demand_max_price_percentage_over_lowest_price: Option<i32>,
    pub bare_metal: Option<BareMetal>,
    pub burstable_performance: Option<BurstablePerformance>,
    pub require_hibernate_support: Option<bool>,
    pub network_interface_count: Option<NetworkInterfaceCountRequest>,
    pub local_storage: Option<LocalStorage>,
    pub local_storage_types: Option<Vec<LocalStorageType>>,
    pub total_local_storage_gb: Option<TotalLocalStorageGbRequest>,
    pub baseline_ebs_bandwidth_mbps: Option<BaselineEbsBandwidthMbpsRequest>,
    pub accelerator_types: Option<Vec<AcceleratorType>>,
    pub accelerator_count: Option<AcceleratorCountRequest>,
    pub accelerator_manufacturers: Option<Vec<AcceleratorManufacturer>>,
    pub accelerator_names: Option<Vec<AcceleratorName>>,
    pub accelerator_total_memory_mi_b: Option<AcceleratorTotalMemoryMiBRequest>,

}

When you specify multiple parameters, you get instance types that satisfy all of the specified parameters. If you specify multiple values for a parameter, you get instance types that satisfy any of the specified values.

Represents requirements for the types of instances that can be launched. You must specify VCpuCount and MemoryMiB, but all other parameters are optional. For more information, see Creating an Auto Scaling group using attribute-based instance type selection in the Amazon EC2 Auto Scaling User Guide.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

v_cpu_count: Option<VCpuCountRequest>

The minimum and maximum number of vCPUs for an instance type.

memory_mi_b: Option<MemoryMiBRequest>

The minimum and maximum instance memory size for an instance type, in MiB.

cpu_manufacturers: Option<Vec<CpuManufacturer>>

Lists which specific CPU manufacturers to include.

For instance types with Intel CPUs, specify intel.

For instance types with AMD CPUs, specify amd.

For instance types with Amazon Web Services CPUs, specify amazon-web-services.

Don't confuse the CPU hardware manufacturer with the CPU hardware architecture. Instances will be launched with a compatible CPU architecture based on the Amazon Machine Image (AMI) that you specify in your launch template.

Default: Any manufacturer

memory_gi_b_per_v_cpu: Option<MemoryGiBPerVCpuRequest>

The minimum and maximum amount of memory per vCPU for an instance type, in GiB.

Default: No minimum or maximum

excluded_instance_types: Option<Vec<String>>

Lists which instance types to exclude. You can use strings with one or more wild cards, represented by an asterisk (*). The following are examples: c5*, m5a.*, r*, *3*.

For example, if you specify c5*, you are excluding the entire C5 instance family, which includes all C5a and C5n instance types. If you specify m5a.*, you are excluding all the M5a instance types, but not the M5n instance types.

Default: No excluded instance types

instance_generations: Option<Vec<InstanceGeneration>>

Indicates whether current or previous generation instance types are included.

For current generation instance types, specify current. The current generation includes EC2 instance types currently recommended for use. This typically includes the latest two to three generations in each instance family. For more information, see Instance types in the Amazon EC2 User Guide for Linux Instances.

For previous generation instance types, specify previous.

Default: Any current or previous generation

spot_max_price_percentage_over_lowest_price: Option<i32>

The price protection threshold for Spot Instances. This is the maximum you’ll pay for a Spot Instance, expressed as a percentage higher than the cheapest M, C, or R instance type with your specified attributes. When Amazon EC2 Auto Scaling selects instance types with your attributes, we will exclude instance types whose price is higher than your threshold. The parameter accepts an integer, which Amazon EC2 Auto Scaling interprets as a percentage. To turn off price protection, specify a high value, such as 999999.

Default: 100

on_demand_max_price_percentage_over_lowest_price: Option<i32>

The price protection threshold for On-Demand Instances. This is the maximum you’ll pay for an On-Demand Instance, expressed as a percentage higher than the cheapest M, C, or R instance type with your specified attributes. When Amazon EC2 Auto Scaling selects instance types with your attributes, we will exclude instance types whose price is higher than your threshold. The parameter accepts an integer, which Amazon EC2 Auto Scaling interprets as a percentage. To turn off price protection, specify a high value, such as 999999.

Default: 20

bare_metal: Option<BareMetal>

Indicates whether bare metal instance types are included, excluded, or required.

Default: excluded

burstable_performance: Option<BurstablePerformance>

Indicates whether burstable performance instance types are included, excluded, or required. For more information, see Burstable performance instances in the Amazon EC2 User Guide for Linux Instances.

Default: excluded

require_hibernate_support: Option<bool>

Indicates whether instance types must provide On-Demand Instance hibernation support.

Default: false

network_interface_count: Option<NetworkInterfaceCountRequest>

The minimum and maximum number of network interfaces for an instance type.

Default: No minimum or maximum

local_storage: Option<LocalStorage>

Indicates whether instance types with instance store volumes are included, excluded, or required. For more information, see Amazon EC2 instance store in the Amazon EC2 User Guide for Linux Instances.

Default: included

local_storage_types: Option<Vec<LocalStorageType>>

Indicates the type of local storage that is required.

For instance types with hard disk drive (HDD) storage, specify hdd.

For instance types with solid state drive (SSD) storage, specify sdd.

Default: Any local storage type

total_local_storage_gb: Option<TotalLocalStorageGbRequest>

The minimum and maximum total local storage size for an instance type, in GB.

Default: No minimum or maximum

baseline_ebs_bandwidth_mbps: Option<BaselineEbsBandwidthMbpsRequest>

The minimum and maximum baseline bandwidth performance for an instance type, in Mbps. For more information, see Amazon EBS–optimized instances in the Amazon EC2 User Guide for Linux Instances.

Default: No minimum or maximum

accelerator_types: Option<Vec<AcceleratorType>>

Lists the accelerator types that must be on an instance type.

For instance types with GPU accelerators, specify gpu.

For instance types with FPGA accelerators, specify fpga.

For instance types with inference accelerators, specify inference.

Default: Any accelerator type

accelerator_count: Option<AcceleratorCountRequest>

The minimum and maximum number of accelerators (GPUs, FPGAs, or Amazon Web Services Inferentia chips) for an instance type.

To exclude accelerator-enabled instance types, set Max to 0.

Default: No minimum or maximum

accelerator_manufacturers: Option<Vec<AcceleratorManufacturer>>

Indicates whether instance types must have accelerators by specific manufacturers.

For instance types with NVIDIA devices, specify nvidia.

For instance types with AMD devices, specify amd.

For instance types with Amazon Web Services devices, specify amazon-web-services.

For instance types with Xilinx devices, specify xilinx.

Default: Any manufacturer

accelerator_names: Option<Vec<AcceleratorName>>

Lists the accelerators that must be on an instance type.

For instance types with NVIDIA A100 GPUs, specify a100.

For instance types with NVIDIA V100 GPUs, specify v100.

For instance types with NVIDIA K80 GPUs, specify k80.

For instance types with NVIDIA T4 GPUs, specify t4.

For instance types with NVIDIA M60 GPUs, specify m60.

For instance types with AMD Radeon Pro V520 GPUs, specify radeon-pro-v520.

For instance types with Xilinx VU9P FPGAs, specify vu9p.

Default: Any accelerator

accelerator_total_memory_mi_b: Option<AcceleratorTotalMemoryMiBRequest>

The minimum and maximum total memory size for the accelerators on an instance type, in MiB.

Default: No minimum or maximum

Implementations

impl InstanceRequirements

pub fn v_cpu_count(&self) -> Option<&VCpuCountRequest>

The minimum and maximum number of vCPUs for an instance type.

pub fn memory_mi_b(&self) -> Option<&MemoryMiBRequest>

The minimum and maximum instance memory size for an instance type, in MiB.

pub fn cpu_manufacturers(&self) -> Option<&[CpuManufacturer]>

Lists which specific CPU manufacturers to include.

For instance types with Intel CPUs, specify intel.

For instance types with AMD CPUs, specify amd.

For instance types with Amazon Web Services CPUs, specify amazon-web-services.

Don't confuse the CPU hardware manufacturer with the CPU hardware architecture. Instances will be launched with a compatible CPU architecture based on the Amazon Machine Image (AMI) that you specify in your launch template.

Default: Any manufacturer

pub fn memory_gi_b_per_v_cpu(&self) -> Option<&MemoryGiBPerVCpuRequest>

The minimum and maximum amount of memory per vCPU for an instance type, in GiB.

Default: No minimum or maximum

pub fn excluded_instance_types(&self) -> Option<&[String]>

Lists which instance types to exclude. You can use strings with one or more wild cards, represented by an asterisk (*). The following are examples: c5*, m5a.*, r*, *3*.

For example, if you specify c5*, you are excluding the entire C5 instance family, which includes all C5a and C5n instance types. If you specify m5a.*, you are excluding all the M5a instance types, but not the M5n instance types.

Default: No excluded instance types

pub fn instance_generations(&self) -> Option<&[InstanceGeneration]>

Indicates whether current or previous generation instance types are included.

For current generation instance types, specify current. The current generation includes EC2 instance types currently recommended for use. This typically includes the latest two to three generations in each instance family. For more information, see Instance types in the Amazon EC2 User Guide for Linux Instances.

For previous generation instance types, specify previous.

Default: Any current or previous generation

pub fn spot_max_price_percentage_over_lowest_price(&self) -> Option<i32>

The price protection threshold for Spot Instances. This is the maximum you’ll pay for a Spot Instance, expressed as a percentage higher than the cheapest M, C, or R instance type with your specified attributes. When Amazon EC2 Auto Scaling selects instance types with your attributes, we will exclude instance types whose price is higher than your threshold. The parameter accepts an integer, which Amazon EC2 Auto Scaling interprets as a percentage. To turn off price protection, specify a high value, such as 999999.

Default: 100

pub fn on_demand_max_price_percentage_over_lowest_price(&self) -> Option<i32>

The price protection threshold for On-Demand Instances. This is the maximum you’ll pay for an On-Demand Instance, expressed as a percentage higher than the cheapest M, C, or R instance type with your specified attributes. When Amazon EC2 Auto Scaling selects instance types with your attributes, we will exclude instance types whose price is higher than your threshold. The parameter accepts an integer, which Amazon EC2 Auto Scaling interprets as a percentage. To turn off price protection, specify a high value, such as 999999.

Default: 20

pub fn bare_metal(&self) -> Option<&BareMetal>

Indicates whether bare metal instance types are included, excluded, or required.

Default: excluded

pub fn burstable_performance(&self) -> Option<&BurstablePerformance>

Indicates whether burstable performance instance types are included, excluded, or required. For more information, see Burstable performance instances in the Amazon EC2 User Guide for Linux Instances.

Default: excluded

pub fn require_hibernate_support(&self) -> Option<bool>

Indicates whether instance types must provide On-Demand Instance hibernation support.

Default: false

pub fn network_interface_count(&self) -> Option<&NetworkInterfaceCountRequest>

The minimum and maximum number of network interfaces for an instance type.

Default: No minimum or maximum

pub fn local_storage(&self) -> Option<&LocalStorage>

Indicates whether instance types with instance store volumes are included, excluded, or required. For more information, see Amazon EC2 instance store in the Amazon EC2 User Guide for Linux Instances.

Default: included

pub fn local_storage_types(&self) -> Option<&[LocalStorageType]>

Indicates the type of local storage that is required.

For instance types with hard disk drive (HDD) storage, specify hdd.

For instance types with solid state drive (SSD) storage, specify sdd.

Default: Any local storage type

pub fn total_local_storage_gb(&self) -> Option<&TotalLocalStorageGbRequest>

The minimum and maximum total local storage size for an instance type, in GB.

Default: No minimum or maximum

pub fn baseline_ebs_bandwidth_mbps(
&self
) -> Option<&BaselineEbsBandwidthMbpsRequest>

The minimum and maximum baseline bandwidth performance for an instance type, in Mbps. For more information, see Amazon EBS–optimized instances in the Amazon EC2 User Guide for Linux Instances.

Default: No minimum or maximum

pub fn accelerator_types(&self) -> Option<&[AcceleratorType]>

Lists the accelerator types that must be on an instance type.

For instance types with GPU accelerators, specify gpu.

For instance types with FPGA accelerators, specify fpga.

For instance types with inference accelerators, specify inference.

Default: Any accelerator type

pub fn accelerator_count(&self) -> Option<&AcceleratorCountRequest>

The minimum and maximum number of accelerators (GPUs, FPGAs, or Amazon Web Services Inferentia chips) for an instance type.

To exclude accelerator-enabled instance types, set Max to 0.

Default: No minimum or maximum

pub fn accelerator_manufacturers(&self) -> Option<&[AcceleratorManufacturer]>

Indicates whether instance types must have accelerators by specific manufacturers.

For instance types with NVIDIA devices, specify nvidia.

For instance types with AMD devices, specify amd.

For instance types with Amazon Web Services devices, specify amazon-web-services.

For instance types with Xilinx devices, specify xilinx.

Default: Any manufacturer

pub fn accelerator_names(&self) -> Option<&[AcceleratorName]>

Lists the accelerators that must be on an instance type.

For instance types with NVIDIA A100 GPUs, specify a100.

For instance types with NVIDIA V100 GPUs, specify v100.

For instance types with NVIDIA K80 GPUs, specify k80.

For instance types with NVIDIA T4 GPUs, specify t4.

For instance types with NVIDIA M60 GPUs, specify m60.

For instance types with AMD Radeon Pro V520 GPUs, specify radeon-pro-v520.

For instance types with Xilinx VU9P FPGAs, specify vu9p.

Default: Any accelerator

pub fn accelerator_total_memory_mi_b(
&self
) -> Option<&AcceleratorTotalMemoryMiBRequest>

The minimum and maximum total memory size for the accelerators on an instance type, in MiB.

Default: No minimum or maximum

impl InstanceRequirements

pub fn builder() -> Builder

Creates a new builder-style object to manufacture InstanceRequirements