Module aws_sdk_autoscaling::operation

Structs

AttachInstances

Attaches one or more EC2 instances to the specified Auto Scaling group.

AttachLoadBalancerTargetGroups

Attaches one or more target groups to the specified Auto Scaling group.

AttachLoadBalancers

BatchDeleteScheduledAction

Deletes one or more scheduled actions for the specified Auto Scaling group.

BatchPutScheduledUpdateGroupAction

Creates or updates one or more scheduled scaling actions for an Auto Scaling group.

CancelInstanceRefresh

Cancels an instance refresh operation in progress. Cancellation does not roll back any

CompleteLifecycleAction

Completes the lifecycle action for the specified token or instance with the specified

CreateAutoScalingGroup

CreateLaunchConfiguration

Creates a launch configuration.

CreateOrUpdateTags

Creates or updates tags for the specified Auto Scaling group.

DeleteAutoScalingGroup

Deletes the specified Auto Scaling group.

DeleteLaunchConfiguration

Deletes the specified launch configuration.

DeleteLifecycleHook

Deletes the specified lifecycle hook.

DeleteNotificationConfiguration

Deletes the specified notification.

DeletePolicy

Deletes the specified scaling policy.

DeleteScheduledAction

Deletes the specified scheduled action.

DeleteTags

Deletes the specified tags.

DeleteWarmPool

Deletes the warm pool for the specified Auto Scaling group.

DescribeAccountLimits

Describes the current Amazon EC2 Auto Scaling resource quotas for your account.

DescribeAdjustmentTypes

Describes the available adjustment types for step scaling and simple scaling

DescribeAutoScalingGroups

Gets information about the Auto Scaling groups in the account and Region.

DescribeAutoScalingInstances

Gets information about the Auto Scaling instances in the account and Region.

DescribeAutoScalingNotificationTypes

Describes the notification types that are supported by Amazon EC2 Auto Scaling.

DescribeInstanceRefreshes

Gets information about the instance refreshes for the specified Auto Scaling group.

DescribeLaunchConfigurations

Gets information about the launch configurations in the account and Region.

DescribeLifecycleHookTypes

Describes the available types of lifecycle hooks.

DescribeLifecycleHooks

Gets information about the lifecycle hooks for the specified Auto Scaling group.

DescribeLoadBalancerTargetGroups

Gets information about the load balancer target groups for the specified Auto Scaling

DescribeLoadBalancers

Gets information about the load balancers for the specified Auto Scaling group.

DescribeMetricCollectionTypes

Describes the available CloudWatch metrics for Amazon EC2 Auto Scaling.

DescribeNotificationConfigurations

Gets information about the Amazon SNS notifications that are configured for one or more

DescribePolicies

Gets information about the scaling policies in the account and Region.

DescribeScalingActivities

Gets information about the scaling activities in the account and Region.

DescribeScalingProcessTypes

Describes the scaling process types for use with the ResumeProcesses

DescribeScheduledActions

Gets information about the scheduled actions that haven't run or that have not reached

DescribeTags

Describes the specified tags.

DescribeTerminationPolicyTypes

Describes the termination policies supported by Amazon EC2 Auto Scaling.

DescribeWarmPool

Gets information about a warm pool and its instances.

DetachInstances

Removes one or more instances from the specified Auto Scaling group.

DetachLoadBalancerTargetGroups

Detaches one or more target groups from the specified Auto Scaling group.

DetachLoadBalancers

Detaches one or more Classic Load Balancers from the specified Auto Scaling group.

DisableMetricsCollection

Disables group metrics for the specified Auto Scaling group.

EnableMetricsCollection

Enables group metrics for the specified Auto Scaling group. For more information, see Monitoring CloudWatch metrics for your Auto Scaling groups and instances in the

EnterStandby

Moves the specified instances into the standby state.

ExecutePolicy

Executes the specified policy. This can be useful for testing the design of your

ExitStandby

Moves the specified instances out of the standby state.

GetPredictiveScalingForecast

Retrieves the forecast data for a predictive scaling policy.

PutLifecycleHook

Creates or updates a lifecycle hook for the specified Auto Scaling group.

PutNotificationConfiguration

Configures an Auto Scaling group to send notifications when specified events take place.

PutScalingPolicy

Creates or updates a scaling policy for an Auto Scaling group. Scaling policies are used to

PutScheduledUpdateGroupAction

Creates or updates a scheduled scaling action for an Auto Scaling group.

PutWarmPool

Creates or updates a warm pool for the specified Auto Scaling group. A warm pool is a pool of

RecordLifecycleActionHeartbeat

Records a heartbeat for the lifecycle action associated with the specified token or

ResumeProcesses

Resumes the specified suspended auto scaling processes, or all suspended process, for

SetDesiredCapacity

Sets the size of the specified Auto Scaling group.

SetInstanceHealth

Sets the health status of the specified instance.

SetInstanceProtection

Updates the instance protection settings of the specified instances. This operation

StartInstanceRefresh

Starts a new instance refresh operation, which triggers a rolling replacement of

SuspendProcesses

Suspends the specified auto scaling processes, or all processes, for the specified

TerminateInstanceInAutoScalingGroup

Terminates the specified instance and optionally adjusts the desired group size. This

UpdateAutoScalingGroup