Struct aws_sdk_autoscaling::input::create_auto_scaling_group_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateAutoScalingGroupInput

Implementations

impl Builder

pub fn auto_scaling_group_name(self, input: impl Into<String>) -> Self

The name of the Auto Scaling group. This name must be unique per Region per account.

pub fn set_auto_scaling_group_name(self, input: Option<String>) -> Self

pub fn launch_configuration_name(self, input: impl Into<String>) -> Self

The name of the launch configuration to use to launch instances.

Conditional: You must specify either a launch template (LaunchTemplate or MixedInstancesPolicy) or a launch configuration (LaunchConfigurationName or InstanceId).

pub fn set_launch_configuration_name(self, input: Option<String>) -> Self

pub fn launch_template(self, input: LaunchTemplateSpecification) -> Self

Parameters used to specify the launch template and version to use to launch instances.

Conditional: You must specify either a launch template (LaunchTemplate or MixedInstancesPolicy) or a launch configuration (LaunchConfigurationName or InstanceId).

The launch template that is specified must be configured for use with an Auto Scaling group. For more information, see Creating a launch template for an Auto Scaling group in the Amazon EC2 Auto Scaling User Guide.

pub fn set_launch_template(
self,
input: Option<LaunchTemplateSpecification>
) -> Self

pub fn mixed_instances_policy(self, input: MixedInstancesPolicy) -> Self

An embedded object that specifies a mixed instances policy. The required properties must be specified. If optional properties are unspecified, their default values are used.

The policy includes properties that not only define the distribution of On-Demand Instances and Spot Instances, the maximum price to pay for Spot Instances, and how the Auto Scaling group allocates instance types to fulfill On-Demand and Spot capacities, but also the properties that specify the instance configuration information—the launch template and instance types. The policy can also include a weight for each instance type and different launch templates for individual instance types. For more information, see Auto Scaling groups with multiple instance types and purchase options in the Amazon EC2 Auto Scaling User Guide.

pub fn set_mixed_instances_policy(
self,
input: Option<MixedInstancesPolicy>
) -> Self

pub fn instance_id(self, input: impl Into<String>) -> Self

The ID of the instance used to base the launch configuration on. If specified, Amazon EC2 Auto Scaling uses the configuration values from the specified instance to create a new launch configuration. To get the instance ID, use the Amazon EC2 DescribeInstances API operation. For more information, see Creating an Auto Scaling group using an EC2 instance in the Amazon EC2 Auto Scaling User Guide.

pub fn set_instance_id(self, input: Option<String>) -> Self

pub fn min_size(self, input: i32) -> Self

The minimum size of the group.

pub fn set_min_size(self, input: Option<i32>) -> Self

pub fn max_size(self, input: i32) -> Self

The maximum size of the group.

With a mixed instances policy that uses instance weighting, Amazon EC2 Auto Scaling may need to go above MaxSize to meet your capacity requirements. In this event, Amazon EC2 Auto Scaling will never go above MaxSize by more than your largest instance weight (weights that define how many units each instance contributes to the desired capacity of the group).

pub fn set_max_size(self, input: Option<i32>) -> Self

pub fn desired_capacity(self, input: i32) -> Self

The desired capacity is the initial capacity of the Auto Scaling group at the time of its creation and the capacity it attempts to maintain. It can scale beyond this capacity if you configure auto scaling. This number must be greater than or equal to the minimum size of the group and less than or equal to the maximum size of the group. If you do not specify a desired capacity, the default is the minimum size of the group.

pub fn set_desired_capacity(self, input: Option<i32>) -> Self

pub fn default_cooldown(self, input: i32) -> Self

The amount of time, in seconds, after a scaling activity completes before another scaling activity can start. The default value is 300. This setting applies when using simple scaling policies, but not when using other scaling policies or scheduled scaling. For more information, see Scaling cooldowns for Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

pub fn set_default_cooldown(self, input: Option<i32>) -> Self

pub fn availability_zones(self, input: impl Into<String>) -> Self

pub fn set_availability_zones(self, input: Option<Vec<String>>) -> Self

pub fn load_balancer_names(self, input: impl Into<String>) -> Self

pub fn set_load_balancer_names(self, input: Option<Vec<String>>) -> Self

pub fn target_group_ar_ns(self, input: impl Into<String>) -> Self

pub fn set_target_group_ar_ns(self, input: Option<Vec<String>>) -> Self

pub fn health_check_type(self, input: impl Into<String>) -> Self

The service to use for the health checks. The valid values are EC2 (default) and ELB. If you configure an Auto Scaling group to use load balancer (ELB) health checks, it considers the instance unhealthy if it fails either the EC2 status checks or the load balancer health checks. For more information, see Health checks for Auto Scaling instances in the Amazon EC2 Auto Scaling User Guide.

pub fn set_health_check_type(self, input: Option<String>) -> Self

pub fn health_check_grace_period(self, input: i32) -> Self

The amount of time, in seconds, that Amazon EC2 Auto Scaling waits before checking the health status of an EC2 instance that has come into service. During this time, any health check failures for the instance are ignored. The default value is 0. For more information, see Health check grace period in the Amazon EC2 Auto Scaling User Guide.

Conditional: Required if you are adding an ELB health check.

pub fn set_health_check_grace_period(self, input: Option<i32>) -> Self

pub fn placement_group(self, input: impl Into<String>) -> Self

The name of an existing placement group into which to launch your instances, if any. A placement group is a logical grouping of instances within a single Availability Zone. You cannot specify multiple Availability Zones and a placement group. For more information, see Placement Groups in the Amazon EC2 User Guide for Linux Instances.

pub fn set_placement_group(self, input: Option<String>) -> Self

pub fn vpc_zone_identifier(self, input: impl Into<String>) -> Self

A comma-separated list of subnet IDs for a virtual private cloud (VPC) where instances in the Auto Scaling group can be created. If you specify VPCZoneIdentifier with AvailabilityZones, the subnets that you specify for this parameter must reside in those Availability Zones.

Conditional: If your account supports EC2-Classic and VPC, this parameter is required to launch instances into a VPC.

pub fn set_vpc_zone_identifier(self, input: Option<String>) -> Self

pub fn termination_policies(self, input: impl Into<String>) -> Self

pub fn set_termination_policies(self, input: Option<Vec<String>>) -> Self

pub fn new_instances_protected_from_scale_in(self, input: bool) -> Self

Indicates whether newly launched instances are protected from termination by Amazon EC2 Auto Scaling when scaling in. For more information about preventing instances from terminating on scale in, see Instance scale-in protection in the Amazon EC2 Auto Scaling User Guide.

pub fn set_new_instances_protected_from_scale_in(
self,
input: Option<bool>
) -> Self

pub fn capacity_rebalance(self, input: bool) -> Self

Indicates whether Capacity Rebalancing is enabled. Otherwise, Capacity Rebalancing is disabled. When you turn on Capacity Rebalancing, Amazon EC2 Auto Scaling attempts to launch a Spot Instance whenever Amazon EC2 notifies that a Spot Instance is at an elevated risk of interruption. After launching a new instance, it then terminates an old instance. For more information, see Amazon EC2 Auto Scaling Capacity Rebalancing in the Amazon EC2 Auto Scaling User Guide.

pub fn set_capacity_rebalance(self, input: Option<bool>) -> Self

pub fn lifecycle_hook_specification_list(
self,
input: impl Into<LifecycleHookSpecification>
) -> Self

pub fn set_lifecycle_hook_specification_list(
self,
input: Option<Vec<LifecycleHookSpecification>>
) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn service_linked_role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the service-linked role that the Auto Scaling group uses to call other Amazon Web Services on your behalf. By default, Amazon EC2 Auto Scaling uses a service-linked role named AWSServiceRoleForAutoScaling, which it creates if it does not exist. For more information, see Service-linked roles in the Amazon EC2 Auto Scaling User Guide.

pub fn set_service_linked_role_arn(self, input: Option<String>) -> Self

pub fn max_instance_lifetime(self, input: i32) -> Self

The maximum amount of time, in seconds, that an instance can be in service. The default is null. If specified, the value must be either 0 or a number equal to or greater than 86,400 seconds (1 day). For more information, see Replacing Auto Scaling instances based on maximum instance lifetime in the Amazon EC2 Auto Scaling User Guide.

pub fn set_max_instance_lifetime(self, input: Option<i32>) -> Self

pub fn context(self, input: impl Into<String>) -> Self

Reserved.

pub fn set_context(self, input: Option<String>) -> Self

pub fn build(self) -> Result<CreateAutoScalingGroupInput, BuildError>

Consumes the builder and constructs a CreateAutoScalingGroupInput